+++ database/migrations/2016_04_23_190808_create_processes_table.php
@@ -13,8 +13,8 @@ class CreateProcessesTable extends Migration
     public function up()
     {
         Schema::create('processes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('child_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('child_id')->unsigned();
             $table->string('desc');
 
             $table->baseActions();
