+++ database/migrations/2018_09_11_132659_add_featured_media_id_to_child.php
@@ -14,7 +14,7 @@ class AddFeaturedMediaIdToChild extends Migration
     public function up()
     {
         Schema::table('children', function (Blueprint $table) {
-            $table->integer('featured_media_id')->nullable()->unsigned();
+            $table->bigInteger('featured_media_id')->nullable()->unsigned();
         });
     }
